% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensimpl_functions.R
\name{batchGenes}
\alias{batchGenes}
\title{Get batch gene information.}
\usage{
batchGenes(ids, species = NULL, version = NULL, includeAll = TRUE,
  extended = FALSE, debug = FALSE)
}
\arguments{
\item{ids}{A list of Ensembl identifiers.}

\item{species}{Either 'Mm' or 'Hs', will default to 'Mm'.}

\item{version}{will default to most recent ensimpl version.}

\item{includeAll}{if TRUE, will give a one to one match}

\item{extended}{TRUE for extra information. NOT CURRENTLY USED}

\item{debug}{Display some debugging information.}
}
\description{
Get batch gene information.
}
\examples{
genes <- batchGenes(c('ENSMUSG00000000001', 'ENSMUSG00000000002'))
}
