% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensimpl_functions.R
\name{batchGenesFile}
\alias{batchGenesFile}
\title{Get batch gene information from a file.}
\usage{
batchGenesFile(fileName, header = FALSE, species = NULL,
  release = NULL, details = FALSE, debug = FALSE)
}
\arguments{
\item{fileName}{A file name with ensembl ids}

\item{header}{TRUE if the file has a header line.}

\item{species}{Either 'Mm' or 'Hs', will default to 'Mm'.}

\item{release}{will default to most recent ensimpl release.}

\item{details}{TRUE for extra information.}

\item{debug}{Display some debugging information.}
}
\value{
By default, with details=FALSE, a tibble will be returned.
  If details=TRUE, a list will be returned with each element being
  the matching "id".
}
\description{
Get batch gene information from a file.
}
