% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensimpl_functions.R
\name{externalDBs}
\alias{externalDBs}
\title{Get the external databases.}
\usage{
externalDBs(species = NULL, release = NULL, debug = FALSE)
}
\arguments{
\item{species}{Either 'Mm' or 'Hs', will default to 'Mm'.}

\item{release}{will default to most recent ensimpl release.}

\item{debug}{Display some debugging information.}
}
\value{
A tibble of external database identifiers.
}
\description{
Get the external databases.
}
\examples{
dbs <- externalDBs()
}
