% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensimpl_functions.R
\name{getGene}
\alias{getGene}
\title{Get gene information.}
\usage{
getGene(id, species = NULL, version = NULL, extended = FALSE,
  debug = FALSE)
}
\arguments{
\item{id}{The Ensembl identifier.}

\item{species}{Either 'Mm' or 'Hs', will default to 'Mm'.}

\item{version}{will default to most recent ensimpl version.}

\item{extended}{TRUE for extra information.}

\item{debug}{Display some debugging information.}
}
\description{
Get gene information.
}
\examples{
gene <- getGene('ENSMUSG00000000001')
}
