% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensimpl_functions.R
\name{releases}
\alias{releases}
\title{Get all the releases information from ensimpl.}
\usage{
releases(debug = FALSE)
}
\arguments{
\item{debug}{Display some debugging information.}
}
\value{
A tibble consisting of species, release and other columns.
}
\description{
Get all the releases information from ensimpl.
}
\examples{
releases()
}
