% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensimpl_functions.R
\name{searchGenes}
\alias{searchGenes}
\title{Search for gene information.}
\usage{
searchGenes(term, species = NULL, release = NULL, debug = FALSE)
}
\arguments{
\item{term}{The search term.}

\item{species}{Either 'Mm' or 'Hs', will default to 'Mm'.}

\item{release}{will default to most recent ensimpl release.}

\item{debug}{Display some debugging information.}
}
\description{
Search for gene information.
}
\examples{
results <- searchGenes('Ace')
}
