% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensimpl_functions.R
\name{versions}
\alias{versions}
\title{Get the version information from ensimpl.}
\usage{
versions(debug = FALSE)
}
\arguments{
\item{debug}{Display some debugging information.}
}
\description{
Get the version information from ensimpl.
}
\examples{
versions()
}
