% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensimpl_functions.R
\name{batchGenesFile}
\alias{batchGenesFile}
\title{Get batch gene information from a file.}
\usage{
batchGenesFile(fileName, header = FALSE, species = NULL,
  version = NULL, includeAll = TRUE, extended = FALSE,
  debug = FALSE)
}
\arguments{
\item{fileName}{A file name with ensembl ids}

\item{species}{Either 'Mm' or 'Hs', will default to 'Mm'.}

\item{version}{will default to most recent ensimpl version.}

\item{includeAll}{if TRUE, will give a one to one match}

\item{extended}{TRUE for extra information. NOT CURRENTLY USED}

\item{debug}{Display some debugging information.}
}
\description{
Get batch gene information from a file.
}
\examples{
genes <- batchGenesFile()
}
